requirements.txt
flask
Pillow
numpy
scikit-learn
